v1.0
This is the first completed version. I don't know if I will be updating this beyond here as it seems to do its job just fine. I have left the lines of debugging in for your own testing and use.
Probably is gross formating but oh well I am not a profesional coder.